+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #435 ByteSources.process() exception for single-entry source
 * #437 ByteSource doesn't close streams in internal methods
 * #451 Decimal equality should work if two decimals only differ in trailing fractional zeros
+* #455 Insert row into dataframe
 
 ## Release 1.1.0